Refactor vector based wfs plugins
