+++ bundles/integration/admin-layerselector/templates/layer/wfslayerSettingsTemplateFooter.html
@@ -1,17 +1,10 @@
 
 
 <% 
-    var showMVTOptions = Oskari.app.getBundleInstanceConfigurationByName('mapfull').conf.plugins
-        .find(function(plugin) {
-            vectorPlugins = [
-                'Oskari.wfsmvt.WfsMvtLayerPlugin', 
-                'Oskari.wfsvector.WfsVectorLayerPlugin'
-            ]
-            return vectorPlugins.indexOf(plugin.id) !== -1;
-        }
-    );
+    var mapModule = Oskari.getSandbox().findRegisteredModuleInstance('MainMapModule');
+    var wfsPlugin = mapModule.getLayerPlugins(model.getLayerType());
 %>
-<div class="add-layer-record" <% if (!showMVTOptions) { %>style="display: none"<% } %>>
+<div class="add-layer-record" <% if (!wfsPlugin.oskariStyleSupport) { %>style="display: none"<% } %>>
     <%
         var styles = model.getMVTStylesWithoutSrcLayer();
         styles = styles ? JSON.stringify(styles) : '';
@@ -24,7 +17,7 @@
             placeholder="<%- instance.getLocalization('admin').mvtWfsOskariStylesDesc %>"><%=styles%></textarea>
     </div>
 </div>
-<div class="add-layer-record" <% if (!showMVTOptions) { %>style="display: none"<% } %>>
+<div class="add-layer-record" <% if (!wfsPlugin.oskariStyleSupport) { %>style="display: none"<% } %>>
     <% var hover = (_.isEmpty(model.getOptions()) || _.isEmpty(model.getOptions().hover)) ? '' : JSON.stringify(model.getOptions().hover); %>
     <div class="add-layer-label-area">
         <label class="add-layer-label" title="<%= instance.getLocalization('admin').mvtHover %>"><%= instance.getLocalization('admin').mvtHover %></label>
@@ -34,7 +27,7 @@
             placeholder="<%- instance.getLocalization('admin').mvtHoverDesc %>"><%=hover%></textarea>
     </div>
 </div>
-<% if (!showMVTOptions) { %>
+<% if (!wfsPlugin.oskariStyleSupport) { %>
 <div class="add-layer-record">
     <div class="add-layer-label-area">
         <label for="add-layer-sld-style" class="add-layer-label"  title="<%= instance.getLocalization('admin').addSldStyleDesc %>"><%= instance.getLocalization('admin').addSldStyle %></label>
